refactor(Board)[OK-89] Add UndoRedoContext, improve FileContext, add KeydownListenerWrapper
