+++ frontend/src/components/board/SlideFileInputs.tsx
@@ -1,11 +1,9 @@
 "use client";
 import React from "react";
-import { useFileContext } from "@/contexts/FileContext";
-import { useMenuData } from "@/contexts/MenuDataContext";
+import { useFile } from "@/contexts/FileContext";
 
 const SlideFileInputs: React.FC = () => {
-  const { handleFileChange, handleLoadImagesFromJson } = useMenuData();
-  const { fileInputRef, fileJSONInputRef } = useFileContext();
+  const { handleFileChange, handleLoadImagesFromJson, fileInputRef, fileJSONInputRef } = useFile();
   return (
     <>
       <input type="file" id="file-input1" className="hidden" ref={fileInputRef} onChange={handleFileChange} />
