+++ frontend/src/contexts/CanvasContext.tsx
@@ -122,7 +122,7 @@ export const CanvasProvider: React.FC<{ children: React.ReactNode }> = ({ childr
 export const useCanvas = () => {
   const context = useContext(CanvasContext);
   if (!context) {
-    throw new Error("useCanvasContext must be used within a CanvasProvider");
+    throw new Error("useCanvas must be used within a CanvasProvider");
   }
   return context;
 };
